+++ apps/examples/blinkid-core-api/src/index.ts
@@ -7,15 +7,24 @@ import {
   ProgressStatusCallback,
 } from "@microblink/blinkid-core";
 
+// Create a progress element to visualize the loading
+const progressElement = document.createElement("div");
+progressElement.style.position = "fixed";
+progressElement.style.top = "20px";
+progressElement.style.left = "20px";
+progressElement.style.padding = "10px";
+progressElement.style.background = "#fff";
+progressElement.style.border = "1px solid #ccc";
+progressElement.style.borderRadius = "4px";
+document.body.appendChild(progressElement);
+
 const progressCallback: ProgressStatusCallback = (progress) => {
-  console.log(progress);
+  progressElement.textContent = `Loading: ${progress.progress}%`;
 };
 
 const blinkIdCore = await loadBlinkIdCore(
   {
     licenseKey: import.meta.env.VITE_LICENCE_KEY,
-    // resourcesLocation: "https://localhost:3020",
-    // wasmVariant: "advanced",
   },
   progressCallback,
 );
@@ -27,3 +36,6 @@ console.log("frameResult", frameResult);
 
 const result = await session.getResult();
 console.log("result", result);
+
+// cleanup memory
+await blinkIdCore.terminate();